+++ src/components/matchups/accordionMatchup/AccordionMatchup.css
@@ -1,0 +1,17 @@
+.table-pagination {
+    display: flex;
+    flex-wrap: nowrap;
+    justify-content: flex-end; /* Adjust this as needed */
+    flex-direction: row;
+    padding-right: 4%;
+  }
+
+.MuiTablePagination-toolbar {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+  }
+
+  /* .MuiTablePagination-selectLabel {
+    padding: 1rem;
+  } */
